feat: Add edit_db_rules script
There are a few yaks to be shaven before we have a good CLI for editing database rules.

This is a simple shell script that uses the grpc API to edit the database rules on a remote
IOx server. It fetches the rules, spawns your default text editor, waits until your editor quits,
and then calls the update API with the results of your local edit.

(if you use emacs, there is a thing called emacsclient which knows how to block until you closed
the buffer in your main persistent emacs)